tutorial/requirement.txt
appdirs==1.4.4
asgiref==3.7.2
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
base58==1.0.3
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
bottle==0.12.25
Brotli==1.0.9
cbor==1.0.0
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
cloud-init==23.2.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.28
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.0.1
djangorestframework==3.14.0
ecdsa==0.18.0b1
entrypoints==0.4
flatbuffers===1.12.1-git20200711.33e2d80-dfsg1-0.6
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
GeoIP==1.3.2
guizero==1.3.0
gyp==0.1
hkdf==0.0.3
html5lib==1.1
httplib2==0.20.2
humanize==0.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
magic-wormhole==0.12.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mnemonic==0.19
mock==4.0.3
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
nest-asyncio==1.5.4
netifaces==0.11.0
norminette==3.3.55
nudatus==0.0.5
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
parso==0.8.1
passlib==1.7.4
pbr==5.8.0
pexpect==4.8.0
pgzero==1.2
pickleshare==0.7.5
Pillow==9.0.1
ply==3.11
prompt-toolkit==3.0.28
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
py-ubjson==0.16.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
pyflakes==2.4.0
pygame==2.1.2
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
python-snappy==0.5.3
pythran==0.10.0
PyTrie==0.4.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
QScintilla==2.11.6
qtconsole==5.2.2
QtPy==2.0.0
requests==2.25.1
scipy==1.8.0
SecretStorage==3.3.1
semver==2.10.2
service-identity==18.1.0
six==1.16.0
sortedcontainers==2.1.0
sos==4.5.6
soupsieve==2.3.1
spake2==0.8
speedtest-cli==2.1.3
sqlparse==0.4.4
ssh-import-id==5.11
supervisor==4.2.1
sympy==1.9
systemd-python==234
terminator==2.1.1
tornado==6.1
tqdm==4.57.0
traitlets==5.1.1
Twisted==22.1.0
txaio==21.2.1
txtorcon==20.0.0
typing_extensions==4.9.0
u-msgpack-python==2.3.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uflash==1.2.4
ufoLib2==0.13.1
ufw==0.36.1
ujson==5.1.0
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
vboxapi==1.0
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
wsaccel==0.6.3
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
